pyproject.toml
[tool.poetry]
name = "stravis"
version = "0.1.0"
description = ""
authors = ["Tim White <tim@timwhite.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
fastapi = "^0.111.0"
fastapi-cli = "^0.0.4"
Jinja2 = "^3.1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
